requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
ntplib==0.4.0
ping3==4.0.8
prompt-toolkit==3.0.43
pyzmq==25.1.1
requests==2.31.0
urllib3==2.2.1
wcwidth==0.2.13
